StoryProof-Project/backend/core/security.py
"""
보안 관련 유틸리티
- JWT 토큰 생성/검증
- 비밀번호 해싱/검증
- 인증 의존성
"""

from datetime import datetime, timedelta
from typing import Optional, Dict, Any
from jose import JWTError, jwt
from passlib.context import CryptContext
from fastapi import Depends, HTTPException, status
from fastapi.security import HTTPBearer, HTTPAuthorizationCredentials

# from backend.core.config import settings


# 비밀번호 해싱 컨텍스트
pwd_context = CryptContext(schemes=["bcrypt"], deprecated="auto")

# HTTP Bearer 토큰 스키마
security = HTTPBearer()


# ===== 비밀번호 관련 함수 =====

def hash_password(password: str) -> str:
    """
    비밀번호 해싱
    
    Args:
        password: 평문 비밀번호
        
    Returns:
        str: 해싱된 비밀번호
    """
    pass


def verify_password(plain_password: str, hashed_password: str) -> bool:
    """
    비밀번호 검증
    
    Args:
        plain_password: 평문 비밀번호
        hashed_password: 해싱된 비밀번호
        
    Returns:
        bool: 비밀번호가 일치하면 True
    """
    pass


# ===== JWT 토큰 관련 함수 =====

def create_access_token(data: Dict[str, Any], expires_delta: Optional[timedelta] = None) -> str:
    """
    액세스 토큰 생성
    
    Args:
        data: 토큰에 포함할 데이터 (user_id, email 등)
        expires_delta: 토큰 만료 시간 (기본값: 30분)
        
    Returns:
        str: JWT 액세스 토큰
    """
    pass


def create_refresh_token(data: Dict[str, Any], expires_delta: Optional[timedelta] = None) -> str:
    """
    리프레시 토큰 생성
    
    Args:
        data: 토큰에 포함할 데이터
        expires_delta: 토큰 만료 시간 (기본값: 7일)
        
    Returns:
        str: JWT 리프레시 토큰
    """
    pass


def decode_token(token: str) -> Dict[str, Any]:
    """
    JWT 토큰 디코딩 및 검증
    
    Args:
        token: JWT 토큰
        
    Returns:
        Dict[str, Any]: 토큰 페이로드
        
    Raises:
        HTTPException: 토큰이 유효하지 않은 경우
    """
    pass


def verify_token(token: str) -> Optional[str]:
    """
    토큰 검증 및 사용자 ID 추출
    
    Args:
        token: JWT 토큰
        
    Returns:
        Optional[str]: 사용자 ID (토큰이 유효하지 않으면 None)
    """
    pass


# ===== 인증 의존성 함수 =====

async def get_current_user_id(
    credentials: HTTPAuthorizationCredentials = Depends(security)
) -> str:
    """
    현재 인증된 사용자 ID 반환 (의존성 주입용)
    
    Args:
        credentials: HTTP Bearer 토큰
        
    Returns:
        str: 사용자 ID
        
    Raises:
        HTTPException: 인증 실패 시
    """
    pass


async def get_current_user(
    user_id: str = Depends(get_current_user_id),
    # db: Session = Depends(get_db)
):
    """
    현재 인증된 사용자 객체 반환 (의존성 주입용)
    
    Args:
        user_id: 사용자 ID
        db: 데이터베이스 세션
        
    Returns:
        User: 사용자 객체
        
    Raises:
        HTTPException: 사용자를 찾을 수 없는 경우
    """
    pass


async def get_optional_user(
    credentials: Optional[HTTPAuthorizationCredentials] = Depends(security)
):
    """
    선택적 인증 (토큰이 없어도 됨)
    
    Args:
        credentials: HTTP Bearer 토큰 (선택)
        
    Returns:
        Optional[User]: 사용자 객체 (인증되지 않으면 None)
    """
    pass


# ===== 권한 검증 함수 =====

def require_admin(current_user = Depends(get_current_user)):
    """
    관리자 권한 필요 (의존성 주입용)
    
    Args:
        current_user: 현재 사용자
        
    Returns:
        User: 사용자 객체
        
    Raises:
        HTTPException: 관리자가 아닌 경우
    """
    pass


def require_verified_email(current_user = Depends(get_current_user)):
    """
    이메일 인증 필요 (의존성 주입용)
    
    Args:
        current_user: 현재 사용자
        
    Returns:
        User: 사용자 객체
        
    Raises:
        HTTPException: 이메일이 인증되지 않은 경우
    """
    pass


# ===== 유틸리티 함수 =====

def generate_verification_token(email: str) -> str:
    """
    이메일 인증 토큰 생성
    
    Args:
        email: 사용자 이메일
        
    Returns:
        str: 인증 토큰
    """
    pass


def verify_verification_token(token: str) -> Optional[str]:
    """
    이메일 인증 토큰 검증
    
    Args:
        token: 인증 토큰
        
    Returns:
        Optional[str]: 이메일 (토큰이 유효하지 않으면 None)
    """
    pass


def generate_password_reset_token(email: str) -> str:
    """
    비밀번호 재설정 토큰 생성
    
    Args:
        email: 사용자 이메일
        
    Returns:
        str: 재설정 토큰
    """
    pass


def verify_password_reset_token(token: str) -> Optional[str]:
    """
    비밀번호 재설정 토큰 검증
    
    Args:
        token: 재설정 토큰
        
    Returns:
        Optional[str]: 이메일 (토큰이 유효하지 않으면 None)
    """
    pass
