StoryProof-Project/backend/requirements.txt
# StoryProof Backend Requirements

## Core Dependencies
fastapi==0.104.1
uvicorn[standard]==0.24.0
pydantic==2.5.0
pydantic-settings==2.1.0

## Database
sqlalchemy==2.0.23
psycopg2-binary==2.9.9
alembic==1.13.0

## Authentication & Security
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-multipart==0.0.6

## AI & LangChain
langchain==0.1.0
openai==1.6.1
chromadb==0.4.18

## File Processing
python-docx==1.1.0
PyPDF2==3.0.1
chardet==5.2.0

## Async & Background Tasks
celery==5.3.4
redis==5.0.1

## Utilities
python-dotenv==1.0.0
